integrate progress bar into transaction confirmation modal
